# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payments": {
      "get": {
        "operationId": "PaymentsController_list",
        "summary": "List the caller’s payments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "summary": "Fetch a payment (owner or admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/payments/{id}/refund": {
      "post": {
        "operationId": "PaymentsController_refund",
        "summary": "Refund a payment (step-up MFA required)",
        "description": "Full or partial. Idempotency-Key guarantees a retry never issues a second refund.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated unique key (UUIDv4 recommended). Replaying the same key with the same payload returns the original response with `Idempotent-Replay: true`. Reusing it with a different payload returns 409. Keys are scoped to user + endpoint and expire after 24h.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5f1d8f4e-6c2b-4a5b-9f1e-3b7c1d2e4a55"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Requires a multi-factor authenticated session."
          },
          "409": {
            "description": "Idempotency-Key reused with a different payload, or an identical request is in flight."
          }
        },
        "tags": [
          "payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "summary": "Payment provider webhook (HMAC verified, replay-safe)",
        "description": "Requires `X-Signature` (hex HMAC-SHA256 of `${timestamp}.${rawBody}`) and `X-Timestamp` (unix seconds, ±5 min). Duplicate eventIds are acknowledged without reprocessing.",
        "parameters": [
          {
            "name": "x-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-timestamp",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Amrutam Telemedicine API",
    "description": "Production-grade telemedicine backend.\n\n**Conventions**\n- Errors follow RFC 7807 `application/problem+json`.\n- Mutating money/slot endpoints require an `Idempotency-Key` header; replays return the original response with `Idempotent-Replay: true`.\n- Lists use opaque cursor pagination (`?cursor=&limit=`).\n- Every response echoes `X-Request-Id` for support correlation.\n- Rate limits are advertised via `RateLimit-*` headers; 429s include `Retry-After`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://api.amrutam.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RefundDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Partial refund amount. Defaults to the full amount."
          },
          "reason": {
            "type": "string",
            "example": "Patient cancelled within the refund window"
          }
        }
      },
      "WebhookDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "example": "evt_01HX8Z0Q"
          },
          "type": {
            "type": "string",
            "example": "payment.captured"
          },
          "providerRef": {
            "type": "string",
            "example": "mock_4b1f..."
          },
          "status": {
            "type": "string",
            "example": "captured"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "eventId",
          "type"
        ]
      },
...
    },
...
  }
}